requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
mysql-connector-python==9.1.0
python-dotenv==1.0.1
pydantic==2.10.3
passlib==1.7.4
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
pyjwt
